refactor placing order chain
